Move static test data to separate file
